requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
python-dotenv==0.13.0
requests==2.22.0
selenium==3.141.0
urllib3==1.25.3
